Address PR review feedback: narrow surface, redesign wire format, fold builder
* dacprivate.h: gate DACSTRESSPRIV_REQUEST_* and DacStressArgGCRefMapRequest
  on #ifdef CDAC_STRESS; redesign as single [in,out] descriptor with
  caller-allocated blob buffer + cbFilled/cbNeeded out fields (drops HR-in-
  payload, drops fixed 252-byte blob).
* cdacstress.cpp: VerifyArgIteratorForMD now uses a 64-byte stack guess
  with heap retry on ERROR_INSUFFICIENT_BUFFER.
* SOSDacImpl handler: alignment-safe ReadUnaligned/WriteUnaligned on the
  request struct; emits standard COM HRESULTs (E_NOTIMPL on decline,
  E_FAIL on encoder throw, ERROR_INSUFFICIENT_BUFFER for two-call).
* ICallingConvention: narrowed to TryComputeArgGCRefMapBlob only, switched
  to standard bool/out-byte[] Try-pattern.
* ArgumentLocation moved to its own file as internal in Contracts.
* CallingConventionGCRefMapBuilder merged into CallingConvention_1
  (net -209 LOC); duplicate GCRefMapToken enum dropped (reuses the
  StackWalkHelpers one used by the decoder).
* GcScanner.cs: reverted to origin/main; PR no longer touches the live
  GC scan path.
* SignatureTypeProvider: reverted GetGenericMethod/TypeParameter from
  virtual to non-virtual (avoiding the breaking change to downstream
  derived types). MethodAndTypeContextProvider now uses interface
  reimplementation (new + explicit re-declaration of
  IRuntimeSignatureTypeProvider) to keep specialized dispatch through
  RuntimeSignatureDecoder.
* TransitionBlock.cs: removed file-wide SA/CA pragma suppressions; fixed
  surfaced warnings (modifier order, trailing whitespace, IsFloatArgument-
  RegisterOffset made virtual, IsTrivialPointerSizedStruct made static).

Verified locally on Windows x64:
  ALLOC+GCREFS+ARGITER stress run, BasicAlloc:    ArgIter 265 pass / 0 fail
  ALLOC+GCREFS+ARGITER stress run, CallSignatures: ArgIter 353 pass / 0 fail
